VAGOV-6046: Rename onclick file to wysiwyg-analytics-setup.
