+++ src/platform/site-wide/index.js
@@ -6,7 +6,7 @@ import '../monitoring/sentry.js';
 import './legacy/menu'; // Used in the footer.
 import './moment-setup';
 import './popups';
-import './onclicks';
+import './wysiwyg-analytics-setup';
 import addMenuListeners from './accessible-menus';
 import startUserNavWidget from './user-nav';
 import startMegaMenuWidget from './mega-menu';
