+++ src/platform/site-wide/wysiwyg-analytics-setup.js
